@@ -1,65 +1,79 @@
 {
   "name": "tms-utkast-mikrofrontend",
   "version": "1.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --port 3000",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "test": "vitest",
     "prepare": "husky install",
-    "lint": "eslint --fix --ext .ts,.tsx,.js,.jsx .",
-    "pretty-quick": "pretty-quick --pattern \"**/*.*(ts|tsx)\""
+    "lint": "eslint --ext .ts,.tsx ."
   },
   "dependencies": {
-    "@navikt/aksel-icons": "^5.6.1",
-    "@navikt/ds-react": "2.9.3",
-    "@navikt/ds-tokens": "2.9.3",
-    "@types/amplitude-js": "8.16.2",
-    "amplitude-js": "8.21.3",
-    "dayjs": "1.11.6",
+    "@navikt/ds-icons": "3.4.3",
+    "@navikt/ds-react": "5.7.3",
+    "@navikt/ds-tokens": "5.7.3",
+    "@types/amplitude-js": "8.16.4",
+    "amplitude-js": "8.21.9",
+    "dayjs": "1.11.10",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-query": "3.39.2"
+    "react-query": "3.39.3"
   },
   "devDependencies": {
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "@types/react": "17.0.33",
-    "@types/react-dom": "17.0.10",
-    "@types/react-modal": "3.13.1",
-    "@typescript-eslint/eslint-plugin": "^5.33.1",
-    "@typescript-eslint/parser": "5.33.1",
-    "@vitejs/plugin-react": "1.0.7",
-    "cross-fetch": "3.1.5",
-    "eslint": "8.22.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-react": "7.30.1",
-    "husky": "7.0.0",
-    "jsdom": "19.0.0",
-    "lint-staged": "13.0.3",
-    "msw": "^0.49.2",
-    "prettier": "2.5.1",
-    "pretty-quick": "3.1.3",
-    "rollup-plugin-import-map": "2.2.2",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "14.5.1",
+    "@types/jest-axe": "3.5.6",
+    "@types/react": "18.2.29",
+    "@types/react-dom": "18.2.14",
+    "@types/react-modal": "3.16.2",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "6.8.0",
+    "@vitejs/plugin-react": "4.1.0",
+    "cross-fetch": "4.0.0",
+    "eslint": "8.51.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "7.33.2",
+    "husky": "8.0.3",
+    "jsdom": "22.1.0",
+    "lint-staged": "15.0.1",
+    "msw": "^1.3.2",
+    "prettier": "3.0.3",
+    "rollup-plugin-import-map": "3.0.0",
     "rollup-plugin-terser": "7.0.2",
-    "typescript": "4.7.2",
-    "vite": "2.9.13",
+    "typescript": "5.2.2",
+    "vite": "4.5.0",
     "vite-plugin-compression": "0.5.1",
-    "vite-plugin-css-injected-by-js": "1.4.0",
+    "vite-plugin-css-injected-by-js": "3.3.0",
     "vite-plugin-mock": "2.9.6",
-    "vitest": "0.26.1",
+    "vitest": "0.34.6",
     "vitest-axe": "0.1.0"
   },
   "prettier": {
     "printWidth": 120
   },
-  "lint-staged": {
-    "*.{js,jsx,ts,tsx}": [
-      "eslint -c .eslintrc.pre-commit.js --fix",
-      "prettier --cache --write"
+  "eslintConfig": {
+    "extends": [
+      "eslint:recommended",
+      "plugin:react/recommended",
+      "plugin:react/jsx-runtime",
+      "plugin:@typescript-eslint/recommended",
+      "plugin:jsx-a11y/recommended"
     ],
-    "*.{css,md,html,json}": [
+    "rules": {
+      "@typescript-eslint/no-unused-vars": "warn",
+      "@typescript-eslint/no-explicit-any": "warn"
+    },
+    "settings": {
+      "react": {
+        "version": "detect"
+      }
+    }
+  },
+  "lint-staged": {
+    "*.{ts,tsx,css,md,html,json}": [
       "prettier --cache --write"
     ]
   },
