@@ -1,11 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import Mikrofrontend from "./Mikrofrontend";
 import "./index.css";
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root") as HTMLElement);
+root.render(
   <React.StrictMode>
     <Mikrofrontend />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
